chore: extrace image realated utils functions into utils section
